Pjax tuning for item add and Item sorting,

+++ controllers/ItemController.php
@@ -60,55 +60,67 @@ class ItemController extends Controller {
 		} else $this->redirect(['person/index']);
 	}
 
-	public function actionDelete($personId, $itemId) {
-		$person = Person::findOne($personId);
-		if ($person) {
+	public function actionDelete($id, $itemId) {
+		$person = Person::findOne($id);
+		$model = Items::findOne($itemId);
+		if ($person && $model) {
 			$session = Yii::$app->session;
-			$model = Items::findOne($itemId);
 
 			$itemSearchModel = new ItemSearch();
-			$itemsDataProvider = $itemSearchModel->search(Yii::$app->request->get(), $personId, 10);
+			$itemsDataProvider = $itemSearchModel->search(Yii::$app->request->get(), $id, 10);
 			$itemModel = new Items();
 
 			try {
 				$model->delete();
 				$session->setFlash('info', Yii::t('app', 'Item') . ' #' . $itemId . ' ' . Yii::t('app', 'deleted'));
-				return $this->renderPartial('//item/itemUpdate', [
-					'person' => $person,
-					'itemsDataProvider' => $itemsDataProvider,
-					'itemSearch' => $itemSearchModel,
-					'itemModel' => $itemModel,
-				]);
 				//	$this->redirect(['person/update','id'=>$personId]); 
 			} catch (\Exception $ex) {
 
 				$session->setFlash('danger', $ex->getMessage());
 			}
+			return $this->renderPartial('//item/itemActionList', [
+				'person' => $person,
+				'itemsDataProvider' => $itemsDataProvider,
+				'itemSearch' => $itemSearchModel,
+				'itemModel' => $itemModel,
+			]);
 		}
 	}
 
-	public function actionUpdate() {
+	public function actionAdd() {
 		$itemModel = new Items();
 
-			// taking data from item add form
+		// taking data from item add form
 
-			if ($itemModel->load($_POST) ) {
-				if ($itemModel->save()) {
-					Yii::$app->session->setFlash('success', 'Added item #' . $itemModel->id . ': '.$itemModel->item);
-				}
-				$person = Person::findOne($itemModel->person_id);
+		if ($itemModel->load($_POST)) {
+			if ($itemModel->save()) {
+				Yii::$app->session->setFlash('success', 'Added item #' . $itemModel->id . ': ' . $itemModel->item);
 			}
+			$person = Person::findOne($itemModel->person_id);
+		}
 
-			$itemSearchModel = new ItemSearch();
-			$itemsDataProvider = $itemSearchModel->search(Yii::$app->request->get(), $person->id, 10);
+		$itemSearchModel = new ItemSearch();
+		$itemsDataProvider = $itemSearchModel->search(Yii::$app->request->get(), $person->id, 10);
 
-			if (Yii::$app->request->isPjax ) {
-				return $this->renderPartial('//item/itemUpdate', [
-					'person' => $person,
-					'itemsDataProvider' => $itemsDataProvider,
-					'itemSearch' => $itemSearchModel,
-					'itemModel' => $itemModel,
-				]);
-			}
+		return $this->renderPartial('//item/itemAdd', [
+			'person' => $person,
+			'itemsDataProvider' => $itemsDataProvider,
+			'itemSearch' => $itemSearchModel,
+			'itemModel' => $itemModel,
+		]);
+	}
+
+	public function actionEditItems($id) {
+
+		$person = Person::findOne($id);
+
+		$itemSearchModel = new ItemSearch();
+		$itemsDataProvider = $itemSearchModel->search(Yii::$app->request->get(), $id, 10);
+
+		return $this->renderPartial('//item/itemActionList', [
+			'person' => $person,
+			'itemsDataProvider' => $itemsDataProvider,
+			'itemSearch' => $itemSearchModel,
+		]);
 	}
 }
